refactor some things around
