+++ runs.py
@@ -11,7 +11,8 @@ from aiohttp.web import Request, Response, HTTPNotFound, HTTPForbidden, HTTPNotI
 
 import aiohttp_jinja2
 
-from helpers.cache_helpers import RunStats
+from cache.cache_helpers import RunStats
+from cache.year_run_stats import update_run_stats
 from sts_profile import get_profile
 from gamedata import FileParser
 from webpage import router
@@ -23,7 +24,6 @@ __all__ = ["get_latest_run"]
 
 _cache: dict[str, RunParser] = {}
 _ts_cache: dict[int, RunParser] = {}
-_run_stats = RunStats()
 
 def get_latest_run(character: str | None, victory: bool | None) -> RunParser:
     _update_cache()
@@ -242,64 +242,6 @@ def _update_cache():
     # that for now, but logging the update time everytime, in case it turns
     # out to be a bottleneck. We only want to actually update new runs.
     logger.info(f"Updated run parser cache in {time.time() - start}s")
-
-def update_run_stats():
-    # we should only have to load this one time this way, after that we can just use the most recent run to update values
-    runs = list(get_profile(1).runs)
-    if not _run_stats.streaks.is_loaded:
-        _run_stats.streaks.all_character_count = 0
-
-        for run in runs[1:]:
-            if run.timestamp.year != 2022:
-                continue
-
-            if run.won:
-                _run_stats.add_win(run.character)
-            else:
-                _run_stats.add_loss(run.character)
-
-            if _run_stats.streaks.is_loaded:
-                continue
-            
-            if _run_stats.streaks.ironclad_count is None and run.character == "Ironclad":
-                _run_stats.streaks.ironclad_count = run.character_streak.streak
-            elif _run_stats.streaks.silent_count is None and run.character == "Silent":
-                _run_stats.streaks.silent_count = run.character_streak.streak
-            elif _run_stats.streaks.defect_count is None and run.character == "Defect":
-                _run_stats.streaks.defect_count = run.character_streak.streak
-            elif _run_stats.streaks.watcher_count is None and run.character == "Watcher":
-                _run_stats.streaks.watcher_count = run.character_streak.streak
-
-    # set the stats from most recent run's rotating streak and character streak
-    last_run = runs[0]
-    if not last_run.modded:
-        print("updating stats")
-        if last_run.won:
-            _run_stats.add_win(last_run.character)
-        else:
-            _run_stats.add_loss(last_run.character)
-        _run_stats.streaks.all_character_count = last_run.rotating_streak.streak
-        match last_run.character:
-            case "Ironclad":
-                _run_stats.streaks.ironclad_count = last_run.character_streak.streak
-            case "Silent":
-                _run_stats.streaks.silent_count = last_run.character_streak.streak
-            case "Defect":
-                _run_stats.streaks.defect_count = last_run.character_streak.streak
-            case "Watcher":
-                _run_stats.streaks.watcher_count = last_run.character_streak.streak
-
-    msg = "A20 Heart kills in 2022: Total: {0} - Ironclad: {1} - Silent: {2} - Defect: {3} - Watcher: {4}"
-    print(msg.format(_run_stats.wins.all_character_count, _run_stats.wins.ironclad_count, _run_stats.wins.silent_count, _run_stats.wins.defect_count, _run_stats.wins.watcher_count))
-
-    msg = "A20 Heart losses in 2022: Total: {0} - Ironclad: {1} - Silent: {2} - Defect: {3} - Watcher: {4}"
-    print(msg.format(_run_stats.losses.all_character_count, _run_stats.losses.ironclad_count, _run_stats.losses.silent_count, _run_stats.losses.defect_count, _run_stats.losses.watcher_count))
-
-    msg = "Current streak: Rotating: {0} - Ironclad: {1} - Silent: {2} - Defect: {3} - Watcher: {4}"
-    print(msg.format(_run_stats.streaks.all_character_count, _run_stats.streaks.ironclad_count, _run_stats.streaks.silent_count, _run_stats.streaks.defect_count, _run_stats.streaks.watcher_count))
-
-def get_run_stats():
-    return _run_stats
 
 @router.get("/runs")
 @aiohttp_jinja2.template("runs_profile.jinja2")
